Correct the MySQL slow query grammar
This version of the parser was built based on the MySQL source in sql/log.cc
(5.1 and 5.5 have the same format).  Moral of the story, only trust the source
code because the initial sample output provided was actually a
Percona-Server-server-51-5.1.58 slow query log and second sample output was a
MySQL 5.1 slow query log.