Add zero copy support to the Heka message and file output (io.write)
One less copy is actually more accurate. This was added to reduce
the memory churn in output plugins for large messages.

For Mozilla telemetry messages the output sandbox memory was reduced from:
raw     1858 KiB to 25 (memory use is fixed in the zero copy case)
framed  3210 KiB to 87 (memory use is still variable since the framing
is created)

For Nginx access log messages there was much less difference
raw        65 KiB to 25
framed     65 KiB to 34

On my Lenovo P50 laptop I saw a 20 increase in throughput using zero copy.
On an AWS c3.2xlarge there was no difference in throughput (I/O appers to
be the gating factor in both cases)