Issue 235 - Allow the backslash to be escaped in the message matcher
An escaped backslash is necessary if a matcher string ends in
backslash otherwise it will simply be passed through as is.
However, this change breaks existing matcher strings with two
consecutive backslashes, they now have to be double escaped.
e.g. "\\\\foo\\bar" would become "\\\\\\\\foo\\bar"
The second backslash does not need escaping since "\b" is not
special in a matcher string.